# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Stockflow API'
  description: 'API REST para inventario y logística de centros de distribución'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Servidor principal'
paths:
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: 'Iniciar sesión con credenciales'
      operationId: ccce227a6532e9aa9166893b0a10946a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
                - device_name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                device_name:
                  type: string
              type: object
      responses:
        200:
          description: 'Autenticación exitosa'
        401:
          description: 'Credenciales inválidas'
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: 'Cerrar sesión y revocar token actual'
      operationId: 16518d98e370a15ebf10146f70441e39
      responses:
        200:
          description: 'Sesión cerrada'
      security:
        -
          bearerAuth: []
  /api/v1/auth/me:
    get:
      tags:
        - Auth
      summary: 'Obtiene el usuario autenticado'
      operationId: a8703cbdf5dcee853c29b34bb2f7c58b
      responses:
        200:
          description: 'Usuario autenticado'
      security:
        -
          bearerAuth: []
  /api/v1/dispatch/orders:
    post:
      tags:
        - Dispatch
      summary: 'Crea orden de despacho'
      operationId: 55a2772047c3e88ce42dd450a734a52c
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - order_reference
                - product_id
                - quantity
              properties:
                order_reference:
                  type: string
                product_id:
                  type: integer
                quantity:
                  type: integer
                  minimum: 1
              type: object
      responses:
        201:
          description: 'Orden creada'
  '/api/v1/dispatch/orders/{dispatchOrderId}/pick-fefo':
    post:
      tags:
        - Dispatch
      summary: 'Ejecuta picking FEFO'
      operationId: a12e92c9405a2d190fcd8bdb32a7084e
      parameters:
        -
          name: dispatchOrderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Picking realizado'
  '/api/v1/dispatch/orders/{dispatchOrderId}/confirm':
    post:
      tags:
        - Dispatch
      summary: 'Confirma despacho y descuenta inventario'
      operationId: cf66c41a6bd8b5aee87e0f337497d97b
      parameters:
        -
          name: dispatchOrderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Despacho confirmado'
  /api/v1/inventory/products:
    get:
      tags:
        - Inventory
      summary: 'Lista productos con paginación por cursor'
      operationId: 0e26126b62eab2ee460e812b592289dd
      parameters:
        -
          name: per_page
          in: query
          schema:
            type: integer
            maximum: 100
            minimum: 1
      responses:
        200:
          description: 'Listado de productos'
    post:
      tags:
        - Inventory
      summary: 'Registra un producto'
      operationId: 733cc3c23682eadf6f9b5dc1cbce58e0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - sku
                - reorder_point
              properties:
                name:
                  type: string
                sku:
                  type: string
                reorder_point:
                  type: integer
                  minimum: 0
              type: object
      responses:
        201:
          description: 'Producto registrado'
  '/api/v1/inventory/products/{productId}':
    get:
      tags:
        - Inventory
      summary: 'Consulta stock de producto'
      operationId: 828f7259295b423d250f0fb4a5e6ae29
      parameters:
        -
          name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Detalle de producto'
  '/api/v1/inventory/products/{productId}/stock':
    patch:
      tags:
        - Inventory
      summary: 'Ajusta stock de un producto'
      operationId: e04c3aa230cfb6bafe8ae5aefbc134c8
      parameters:
        -
          name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - quantity
                - operation
                - reason
              properties:
                quantity:
                  type: integer
                  minimum: 1
                operation:
                  type: string
                  enum:
                    - add
                    - subtract
                reason:
                  type: string
              type: object
      responses:
        200:
          description: 'Stock ajustado'
  /api/v1/inventory/products/below-reorder-point:
    get:
      tags:
        - Inventory
      summary: 'Lista productos bajo punto de reorden'
      operationId: b4b9c061136e8526d5d63927158d890b
      responses:
        200:
          description: 'Listado filtrado'
  /api/v1/receiving/receipts:
    post:
      tags:
        - Receiving
      summary: 'Registra una recepción'
      operationId: 7ab5152eb606bcd529a04b466dba8300
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - supplier_reference
                - received_at
              properties:
                supplier_reference:
                  type: string
                received_at:
                  type: string
                  format: date-time
                notes:
                  type: string
              type: object
      responses:
        201:
          description: 'Recepción creada'
  '/api/v1/receiving/receipts/{receiptId}/lots':
    post:
      tags:
        - Receiving
      summary: 'Asocia lote a recepción'
      operationId: 4649a95b106c7f321c4c9c53f73a7511
      parameters:
        -
          name: receiptId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
                - lot_number
                - expires_at
                - quantity
              properties:
                product_id:
                  type: integer
                lot_number:
                  type: string
                expires_at:
                  type: string
                  format: date
                quantity:
                  type: integer
                location_id:
                  type: integer
              type: object
      responses:
        201:
          description: 'Lote asociado y mercancía recibida'
  /api/v1/reporting/inventory-metrics:
    get:
      tags:
        - Reporting
      summary: 'Obtiene tabla materializada de métricas de inventario'
      operationId: 3e115023c63aa286d8bbfd7562d71df3
      responses:
        200:
          description: 'Métricas de inventario'
  /api/v1/reporting/summary:
    get:
      tags:
        - Reporting
      summary: 'Obtiene indicadores agregados de reporting'
      operationId: d26652dd660d77a4e1b8f76b9f401d1d
      responses:
        200:
          description: 'Resumen de indicadores'
  /api/v1/reporting/inventory-metrics/export/pdf:
    get:
      tags:
        - Reporting
      summary: 'Exporta métricas de inventario en PDF'
      operationId: 0662ca284362255d36db313371e53706
      responses:
        200:
          description: 'Archivo PDF'
  /api/v1/reporting/inventory-metrics/export/excel:
    get:
      tags:
        - Reporting
      summary: 'Exporta métricas de inventario en Excel'
      operationId: 17430feec35bd68d5a265eb48d036202
      responses:
        200:
          description: 'Archivo XLSX'
components:
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: Token
      scheme: bearer
tags:
  -
    name: Auth
    description: Auth
  -
    name: Dispatch
    description: Dispatch
  -
    name: Inventory
    description: Inventory
  -
    name: Receiving
    description: Receiving
  -
    name: Reporting
    description: Reporting